% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{Calculate Fibonacci Numbers}
\usage{
fibonacci(N)
}
\arguments{
\item{N}{The number of Fibonacci numbers to calculate.}
}
\value{
A numeric vector containing the first N Fibonacci numbers.
}
\description{
This function calculates the first N Fibonacci numbers.
}
