% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{Fibonacci sequence}
\usage{
fibonacci(n)
}
\arguments{
\item{n}{A number}
}
\value{
The digit in the sequence
}
\description{
Fibonacci sequence
}
\examples{
fibonacci(5)
fibonacci(12)
}
